+++ components/icons/HomeIcon.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+export const HomeIcon: React.FC<React.SVGProps<SVGSVGElement>> = (props) => (
+  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" {...props}>
+    <path strokeLinecap="round" strokeLinejoin="round" d="M2.25 12l8.954-8.955a.75.75 0 011.06 0l8.955 8.955M3 10.5v.75a3 3 0 003 3h1.5v3a.75.75 0 00.75.75h3a.75.75 0 00.75-.75v-3H18a3 3 0 003-3v-.75" />
+  </svg>
+);